+++ src/main/java/ujin/UjinException.java
@@ -1,4 +1,6 @@
-class UjinException extends Exception {
+package ujin;
+
+public class UjinException extends Exception {
     public UjinException() {}
     public UjinException(String message) {
         super(message);
