+++ src/main/java/ujin/command/AddCommand.java
@@ -1,3 +1,8 @@
+package ujin.command;
+
+import ujin.task.*;
+import ujin.ui.*;
+
 public class AddCommand extends Command {
 
     Task task;
