+++ src/main/java/ujin/command/Command.java
@@ -1,3 +1,8 @@
+package ujin.command;
+
+import ujin.task.*;
+import ujin.ui.*;
+
 public abstract class Command {
 
     public boolean isExit = false;
