+++ src/main/java/ujin/command/DeleteCommand.java
@@ -1,3 +1,8 @@
+package ujin.command;
+
+import ujin.task.*;
+import ujin.ui.*;
+
 public class DeleteCommand extends Command {
 
     private final int index;
