+++ src/main/java/ujin/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package ujin.command;
+
+import ujin.ui.*;
+import ujin.task.*;
+import ujin.helper.*;
+
 public class ExitCommand extends Command {
 
     @Override
