+++ src/main/java/ujin/helper/TaskProcessor.java
@@ -1,7 +1,11 @@
+package ujin.helper;
+
 import java.io.*;
 import java.nio.file.*;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
+import ujin.task.*;
+
 
 public class TaskProcessor {
 
@@ -67,22 +71,22 @@ public class TaskProcessor {
             text += "E";
         }
         text += " | ";
-        text += (task.isDone ? "1" : "0");
+        text += (task.isDone() ? "1" : "0");
         text += " | ";
         if (task instanceof Todo) {
-            text += task.description;
+            text += task.description();
         }
         if (task instanceof Deadline) {
-            text += task.description;
+            text += task.description();
             text += " | ";
-            text += ((Deadline) task).by.format(formatter);
+            text += ((Deadline) task).by().format(formatter);
         }
         if (task instanceof Event) {
-            text += task.description;
+            text += task.description();
             text += " | ";
-            text += ((Event) task).start.format(formatter);
+            text += ((Event) task).start().format(formatter);
             text += " | ";
-            text += ((Event) task).end.format(formatter);
+            text += ((Event) task).end().format(formatter);
         }
         return text;
     }
