+++ src/main/java/ujin/task/Todo.java
@@ -1,3 +1,5 @@
+package ujin.task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
